+++ validationframework-swing/src/main/java/com/google/code/validationframework/swing/property/JComponentToolTipTextProperty.java
@@ -1,0 +1,65 @@
+/*
+ * Copyright (c) 2014, Patrick Moawad
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright notice, this
+ *    list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package com.google.code.validationframework.swing.property;
+
+import javax.swing.JComponent;
+
+/**
+ * Readable/writable property representing the tooltip text client property of a {@see JComponent}.
+ * <p/>
+ * It is possible to control the tooltip text of the component by setting the value of this property or by calling the
+ * {@link JComponent#setToolTipText(String)} method of that component.
+ * <p/>
+ * If the value of this property is set to null, the tooltip will be emptied.
+ *
+ * @see JComponent#getToolTipText()
+ * @see JComponent#setToolTipText(String)
+ */
+public class JComponentToolTipTextProperty extends AbstractComponentProperty<JComponent, String> {
+
+    /**
+     * @see AbstractComponentProperty#AbstractComponentProperty(java.awt.Component, String)
+     */
+    public JComponentToolTipTextProperty(JComponent component) {
+        super(component, JComponent.TOOL_TIP_TEXT_KEY);
+    }
+
+    /**
+     * @see AbstractComponentProperty#getPropertyValueFromComponent()
+     */
+    @Override
+    protected String getPropertyValueFromComponent() {
+        return component.getToolTipText();
+    }
+
+    /**
+     * @see AbstractComponentProperty#setPropertyValueToComponent(Object)
+     */
+    @Override
+    protected void setPropertyValueToComponent(String value) {
+        component.setToolTipText(value);
+    }
+}