+++ CHANGES.md
@@ -1,5 +1,6 @@
 # Version 3.2.1 (to be released)
 
+* Added tooltip text property for JComponent
 * Added convenient comparison method in ValueUtils
 
 # Version 3.1.1